[frontend] WIP Refacto Users view to Datatable (#3343)
