+++ opencti-platform/opencti-front/src/private/components/settings/users/UsersLines.tsx
@@ -53,83 +53,3 @@ const UsersLines: React.FC<UsersLinesProps> = (props) => {
     />
   );
 };
-
-export const usersLinesQuery = graphql`
-  query UsersLinesPaginationQuery(
-    $search: String
-    $count: Int!
-    $cursor: ID
-    $orderBy: UsersOrdering
-    $orderMode: OrderingMode
-  ) {
-    ...UsersLines_data
-      @arguments(
-        search: $search
-        count: $count
-        cursor: $cursor
-        orderBy: $orderBy
-        orderMode: $orderMode
-      )
-  }
-`;
-
-export default createPaginationContainer(
-  UsersLines,
-  {
-    data: graphql`
-      fragment UsersLines_data on Query
-      @argumentDefinitions(
-        search: { type: "String" }
-        count: { type: "Int", defaultValue: 25 }
-        cursor: { type: "ID" }
-        orderBy: { type: "UsersOrdering", defaultValue: name }
-        orderMode: { type: "OrderingMode", defaultValue: asc }
-      ) {
-        users(
-          search: $search
-          first: $count
-          after: $cursor
-          orderBy: $orderBy
-          orderMode: $orderMode
-        ) @connection(key: "Pagination_users") {
-          edges {
-            node {
-              id
-              name
-              firstname
-              lastname
-              ...UserLine_node
-            }
-          }
-          pageInfo {
-            endCursor
-            hasNextPage
-            globalCount
-          }
-        }
-      }
-    `,
-  },
-  {
-    direction: 'forward',
-    getConnectionFromProps(props) {
-      return props.data && props.data.users;
-    },
-    getFragmentVariables(prevVars, totalCount) {
-      return {
-        ...prevVars,
-        count: totalCount,
-      };
-    },
-    getVariables(props, { count, cursor }, fragmentVariables) {
-      return {
-        search: fragmentVariables.search,
-        count,
-        cursor,
-        orderBy: fragmentVariables.orderBy,
-        orderMode: fragmentVariables.orderMode,
-      };
-    },
-    query: usersLinesQuery,
-  },
-);
